1.8.2 (Big Update)
Added Koikatsu Sunshine Compendium to the fray readme will be updated in a bit as well to accommodate this change
Added a list of current plugins within it
Added a bunch of additional plugins to the main list that were missing in previous entries. some slight typo fixes and that would be all for now.